Try running at `document-idle` to initialise later in the page load
The script was running too early, and the content of the element wasn't
populated before it tried to set the title. Try running at
`document-idle` to see if that fixes it. It doesn't. Looks like we're
going to have to use a mutation observer. We would have needed to
anyway, though, because the text can change while the page is open, and
we want the tooltip to update accordingly.